F01: Add implementation of TypeService
