+++ src/main/java/com/carto/sn/service/TypeService.java
@@ -1,5 +1,13 @@
 package com.carto.sn.service;
 
-public interface TypeService {
+import java.util.List;
 
+import com.carto.sn.entities.Type;
+
+public interface TypeService {
+	public Type ajoutType(String type, String couleur);
+	public void supprimerType(Long idType);
+	public List <Type> tousLesTypes();
+	public Type findByNomType(String nomType);
+	public Type findByIdType(Long idType);
 }
